+++ src/command_modules/azure-cli-consumption/azure/cli/command_modules/consumption/_params.py
@@ -5,8 +5,9 @@
 
 # pylint: disable=line-too-long
 # pylint: disable=too-many-statements
-from ._validators import (get_datetime_type,
-                          get_decimal_type)
+from ._validators import (datetime_type,
+                          decimal_type)
+from azure.cli.core.commands.parameters import get_enum_type
 
 
 def load_arguments(self, _):
@@ -14,14 +15,14 @@ def load_arguments(self, _):
         c.argument('top', options_list=['--top', '-t'], type=int, help='Maximum number of items to return. Value range: 1-1000.')
         c.argument('include_additional_properties', options_list=['--include-additional-properties', '-a'], action='store_true', help='Include additional properties in the usages.')
         c.argument('include_meter_details', options_list=['--include-meter-details', '-m'], action='store_true', help='Include meter details in the usages.')
-        c.argument('start_date', options_list=['--start-date', '-s'], type=get_datetime_type(), help='Start date (YYYY-MM-DD in UTC). If specified, also requires --end-date.')
-        c.argument('end_date', options_list=['--end-date', '-e'], type=get_datetime_type(), help='End date (YYYY-MM-DD in UTC). If specified, also requires --start-date.')
+        c.argument('start_date', options_list=['--start-date', '-s'], type=datetime_type, help='Start date (YYYY-MM-DD in UTC). If specified, also requires --end-date.')
+        c.argument('end_date', options_list=['--end-date', '-e'], type=datetime_type, help='End date (YYYY-MM-DD in UTC). If specified, also requires --start-date.')
         c.argument('billing_period_name', options_list=['--billing-period-name', '-p'], help='Name of the billing period to get the usage details that associate with.')
 
     with self.argument_context('consumption reservation') as rs:
         rs.argument('reservation_order_id', options_list='--reservation-order-id', help='Reservation order id.')
-        rs.argument('start_date', options_list=['--start-date', '-s'], type=get_datetime_type(), help='Start date (YYYY-MM-DD in UTC). Only needed for daily grain and if specified, also requires --end-date.')
-        rs.argument('end_date', options_list=['--end-date', '-e'], type=get_datetime_type(), help='End date (YYYY-MM-DD in UTC). Only needed for daily grain and if specified, also requires --start-date.')
+        rs.argument('start_date', options_list=['--start-date', '-s'], type=datetime_type, help='Start date (YYYY-MM-DD in UTC). Only needed for daily grain and if specified, also requires --end-date.')
+        rs.argument('end_date', options_list=['--end-date', '-e'], type=datetime_type, help='End date (YYYY-MM-DD in UTC). Only needed for daily grain and if specified, also requires --start-date.')
         rs.argument('reservation_id', options_list='--reservation-id', help='Reservation id.')
 
     with self.argument_context('consumption reservation summary list') as rs:
@@ -34,17 +35,17 @@ def load_arguments(self, _):
     with self.argument_context('consumption marketplace list') as cmp:
         cmp.argument('billing_period_name', options_list=['--billing-period-name', '-p'], help='Name of the billing period to get the marketplace.')
         cmp.argument('top', options_list=['--top', '-t'], type=int, help='Maximum number of items to return. Value range: 1-1000.')
-        cmp.argument('start_date', options_list=['--start-date', '-s'], type=get_datetime_type(), help='Start date (YYYY-MM-DD in UTC). If specified, also requires --end-date.')
-        cmp.argument('end_date', options_list=['--end-date', '-e'], type=get_datetime_type(), help='End date (YYYY-MM-DD in UTC). If specified, also requires --start-date.')
+        cmp.argument('start_date', options_list=['--start-date', '-s'], type=datetime_type, help='Start date (YYYY-MM-DD in UTC). If specified, also requires --end-date.')
+        cmp.argument('end_date', options_list=['--end-date', '-e'], type=datetime_type, help='End date (YYYY-MM-DD in UTC). If specified, also requires --start-date.')
 
     with self.argument_context('consumption budget') as cb:
         cb.argument('resource_group_name', options_list='--resource-group-name', type=str, help='Resource group of the budget.')
         cb.argument('budget_name', options_list='--budget-name', type=str, help='Name of a budget.')
-        cb.argument('category', options_list=['--category', '-c'], type=str, help='Category of the budget can be cost or usage.')
-        cb.argument('amount', options_list=['--amount', '-a'], type=get_decimal_type(), help='Amount of a budget.')
-        cb.argument('time_grain', options_list='--time-grain', type=str, help='Time grain of the budget can be monthly, quarterly, or annually.')
-        cb.argument('start_date', options_list=['--start-date', '-s'], type=get_datetime_type(), help='Start date (YYYY-MM-DD in UTC) of time period of a budget.')
-        cb.argument('end_date', options_list=['--end-date', '-e'], type=get_datetime_type(), help='End date (YYYY-MM-DD in UTC) of time period of a budget.')
+        cb.argument('category', options_list='--category', arg_type=get_enum_type(['cost', 'usage']), help='Category of the budget can be cost or usage.')
+        cb.argument('amount', options_list='--amount', type=decimal_type, help='Amount of a budget.')
+        cb.argument('time_grain', options_list='--time-grain', arg_type=get_enum_type(['monthly', 'quarterly', 'annually']), help='Time grain of the budget can be monthly, quarterly, or annually.')
+        cb.argument('start_date', options_list=['--start-date', '-s'], type=datetime_type, help='Start date (YYYY-MM-DD in UTC) of time period of a budget.')
+        cb.argument('end_date', options_list=['--end-date', '-e'], type=datetime_type, help='End date (YYYY-MM-DD in UTC) of time period of a budget.')
         cb.argument('resource_groups', options_list='--resource-groups', nargs='+', help='Space-separated list of resource groups to filter on.')
         cb.argument('resources', options_list='--resources', nargs='+', help='Space-separated list of resource instances to filter on.')
         cb.argument('meters', options_list='--meters', nargs='+', help='Space-separated list of meters to filter on. Required if category is usage.')