+++ src/command_modules/azure-cli-consumption/azure/cli/command_modules/consumption/_validators.py
@@ -8,30 +8,24 @@ from decimal import Decimal
 from azure.cli.core.util import CLIError
 
 
-def get_datetime_type():
+def datetime_type(string):
     """ Validates UTC datetime. Examples of accepted forms:
     2017-12-31T01:11:59Z,2017-12-31T01:11Z or 2017-12-31T01Z or 2017-12-31 """
-    def datetime_type(string):
-        """ Validates UTC datetime. Examples of accepted forms:
-        2017-12-31T01:11:59Z,2017-12-31T01:11Z or 2017-12-31T01Z or 2017-12-31 """
-        accepted_date_formats = ['%Y-%m-%dT%H:%M:%SZ', '%Y-%m-%dT%H:%MZ',
-                                 '%Y-%m-%dT%HZ', '%Y-%m-%d']
-        for form in accepted_date_formats:
-            try:
-                return datetime.strptime(string, form)
-            except ValueError:
-                continue
-        raise ValueError("Input '{}' not valid. Valid example: 2017-02-11T23:59:59Z".format(string))
-    return datetime_type
-
-
-def get_decimal_type():
-    def decimal_type(string):
+    accepted_date_formats = ['%Y-%m-%dT%H:%M:%SZ', '%Y-%m-%dT%H:%MZ',
+                                '%Y-%m-%dT%HZ', '%Y-%m-%d']
+    for form in accepted_date_formats:
         try:
-            return Decimal(string)
+            return datetime.strptime(string, form)
         except ValueError:
-            raise ValueError("the value passed cannot be converted to decimal")
-    return decimal_type
+            continue
+    raise ValueError("Input '{}' not valid. Valid example: 2017-02-11T23:59:59Z".format(string))
+
+
+def decimal_type(string):
+    try:
+        return Decimal(string)
+    except ValueError:
+        raise ValueError("the value passed cannot be converted to decimal")
 
 
 def validate_both_start_end_dates(namespace):
@@ -53,12 +47,12 @@ def validate_budget_parameters(namespace):
     """lowercase the time grain for comparison"""
     budgetcategory = namespace.category.lower()
     if budgetcategory != 'cost' and budgetcategory != 'usage':
-        raise CLIError("usage error: --parameters.category must be set to either cost or usage")
+        raise CLIError("usage error: --category must be set to either cost or usage")
 
     time_grain = namespace.time_grain.lower().strip()
 
     if time_grain != 'annually' and time_grain != 'quarterly' and time_grain != 'monthly':
-        raise CLIError("usage error: --parameters.time_grain can be 'Annually', 'Quarterly', or 'Monthly'.")
+        raise CLIError("usage error: --time_grain can be 'Annually', 'Quarterly', or 'Monthly'.")
 
     if namespace.amount < 0:
-        raise CLIError("usage error: --parameters.amount must be greater than 0")
+        raise CLIError("usage error: --amount must be greater than 0")